J.J.P. Stewart:
A fault in subroutine "empiri" that caused random characters to be printed has been corrected.
Zero out the six trivial internal coordinates in the output of SADDLE calculations.
Allow GEO_DAT and GEO_REF filenames to include keywords.
Output to screen information on DRC/IRC.
In jobs that use PDB data, allow either GEO_REF or the data-set to not have PDB data, but use the PDB data from the other set.
Improve diagnostic messages when GEO_REF and the data-set have different empirical formulae.
When HTML is used, and the job writes two HTML files, allow the atomic partial charges to be printed.
In SADDLE (react1.F90) terminate when the distance is 0.1 A, was 0.01A and that caused an error in several cases.
In geo_ref.F90, the test for bias had been corrupted.  This is now corrected.
Keyword METAL on its own was not being correctly recognized.
A minor formatting fault in keyword OPEN corrected.
Error message for C.I. expanded to describe the fault.
A major fault that prevented state symbols for "d" orbital systems from being printed has been corrected.
The alternative form of quotation marks '“' is now allowed as an alternative to the standard for, '"'.
In a PATH-type job, hystereses in UHF calculations that introduced artifacts has been corrected.