J.J.P. Stewart:
GEOCHK: If a ligand that has a S-S bond is present, do NOT lyse the bond when generating PDB labels.
GEOCHK: Do NOT calculate charge on system unless MOZYME is used.
GEOCHK: Numbering of salt-bridges was incorrect, fault corrected.
GEOCHK and RUN_MOPAC: A bug in the PDB format of atoms that are not in a protein and not in a ligand has been fixed. (Affects isolated hydrogen atoms.)
GEO_REF: Suppress error message preventing a PDB file from being used (the new file is re-named so there is no conflict.)
GEO_REF: If both the system defined by GEO_REF and the system defined by GEO_DAT do not have PDB data, add PDB data.
FORCE: Minor improvement in format - hide unused parentheses when a FORCETS subset is used.
MOLDAT and WRTKEY: Minor improvements in formats.
MOLDAT and WRTKEY: Output better error message when OPEN is incorrect.
CHARST: Increase maximum number of particles in a microstate from a minimum of 3 to a minimum of 6. (This allows for C.I. of 15876 microstates, the maximum possible; the previous practical limit was ca. 7000.)
WRITMO and PDBOUT: If appropriate, write Heat of Formation to HTML file.
LIGAND: Print XENO ligand-names on the first, not the second, call to identify ligands.
ITER: For paths and similar jobs, don't reset the UHF density matrix after the first SCF.
READMO: The largest C.I. allowed has been increased from 10000 states to 16000 states.
READMO: An error message when INT is used has been re-written to make it more informative.
WRTKEY: Fix format bug in METAL.
EXCITE: Two unnecessary de-allocates removed.
DRC: If, in a RESTART from a FORCE calculation, IRC is used instead of IRC=n, print an error message.
DENSITY_FOR_GPU: BUGFIX - Construction of density matrix in UHF OPEN was faulty.